server/requirements.txt
alembic==1.3.2
attrs==19.3.0
certifi==2019.11.28
Click==7.0
coverage==5.0.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
